feat : mainTemplate에 Header 컴포넌트 추가
