+++ src/components/templeit/mainTemplate.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { BoxProps } from '@chakra-ui/react';
 
 import ReservationList from '../organisms/ReservationList';
+import Header from '../organisms/Header';
 
 interface Props {
   ListStyle?: BoxProps;
@@ -10,7 +11,12 @@ interface Props {
 }
 
 const MainTemplate: React.FC<Props> = ({ ListStyle, ItemStyle }) => {
-  return <ReservationList ListStyle={ListStyle} ItemStyle={ItemStyle} />;
+  return (
+    <>
+      <Header />
+      <ReservationList ListStyle={ListStyle} ItemStyle={ItemStyle} />;
+    </>
+  );
 };
 
 export default MainTemplate;
